requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==3.0.3
django-celery-beat==1.6.0
django-celery-results==1.2.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-jet-new==2.2.8
django-rest-knox==4.1.0
django-timezone-field==4.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
drf-yasg==1.17.0
et-xmlfile==1.0.1
feedparser==5.2.1
flatlib==0.2.1
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
kombu==4.6.7
Markdown==3.2
MarkupSafe==1.1.1
numpy==1.18.4
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==20.1
Pillow==7.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycountry==19.8.18
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.6
pyswisseph==2.0.0.post2
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.7
ruamel.yaml.clib==0.2.0
sentry-sdk==0.14.4
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
timezonefinder==4.4.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
vine==1.3.0
zipp==2.2.0
